requirements.txt
pygame-ce>=2.5.0
requests>=2.31.0
pytest>=7.0.0
